Export the icu object if module is defined (as in a Node environment).
Summary: The icu object is not explicitly exported if `module` exists. This updates it to allow this to happen (making it possible to get the icu object in a Node.js environment).

Test Plan: This is used together with my upcoming diff to improve the exports of some other modules.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D26967