[NODE_TESTS 2/20] Add in module.exports to icu.en.js to make testing easier.
Summary: This has no impact in most cases but when loading this file in Node.js (such as in the new Node.js test runner) it will properly export the icu object and become usable.

Test Plan: Hard to test without the Node.js test runner! Should be a safe change, though.

Reviewers: kevinb, jeff

Reviewed By: kevinb, jeff

Differential Revision: https://phabricator.khanacademy.org/D42690